@@ -1,74 +1,70 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <LangVersion>preview</LangVersion>
-    <PackAsTool>true</PackAsTool>
-    <IsTool>true</IsTool>
-    <ToolCommandName>perfx</ToolCommandName>
-    <PackageOutputPath>./bin</PackageOutputPath>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Authors>Vamsi TP</Authors>
-    <Title>Perfx</Title>
-    <PackageDescription>Azure API Performance benchmarking tool</PackageDescription>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/vamsitp/perfx</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/vamsitp/perfx</RepositoryUrl>
-    <PackageTags>Azure;API;Perf;Performance;Benchmark;Throughput;SLA</PackageTags>
-    <Company>Microsoft</Company>
-    <Version>1.3.5</Version>
-    <StartupObject>Perfx.Program</StartupObject>
-    <ApplicationIcon>logo.ico</ApplicationIcon>
-    <Deterministic>false</Deterministic>
-    <NoWarn>NU1701</NoWarn>
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<!--<TargetFramework>net8.0-windows10.0.20348.0</TargetFramework>-->
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+		<PackAsTool>true</PackAsTool>
+		<ToolCommandName>hi</ToolCommandName>
+		<PackageOutputPath>bin</PackageOutputPath>
+		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+		<LangVersion>latest</LangVersion>
+		<Authors>Vamsi TP</Authors>
+		<Title>Perfx</Title>
+		<PackageDescription>Azure API Performance benchmarking tool</PackageDescription>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<PackageProjectUrl>https://github.com/vamsitp/perfx</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/vamsitp/perfx</RepositoryUrl>
+		<PackageTags>Azure;API;Perf;Performance;Benchmark;Throughput;SLA</PackageTags>
+		<Company>Vamsi TP</Company>
+		<Version>2.0.0</Version>
+		<PublishRepositoryUrl>true</PublishRepositoryUrl>
+		<IncludeSymbols>true</IncludeSymbols>
+		<SymbolPackageFormat>snupkg</SymbolPackageFormat>
+		<ApplicationIcon>logo.ico</ApplicationIcon>
+		<NoWarn>NU1701</NoWarn>
+		<!--<SupportedOSPlatformVersion>10.0.17763.0</SupportedOSPlatformVersion>-->
+		<StartupObject>Perfx.Program</StartupObject>
+	</PropertyGroup>
 
-    <PublishRepositoryUrl>true</PublishRepositoryUrl>
-    <IncludeSymbols>true</IncludeSymbols>
-    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-  </PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="Alba.CsConsoleFormat.ColorfulConsole-NoXaml" Version="1.0.0" />
+		<PackageReference Include="ByteSize" Version="2.1.1" />
+		<PackageReference Include="ClosedXML" Version="0.102.2" />
+		<PackageReference Include="ColoredConsole" Version="1.0.0" />
+		<PackageReference Include="CsvHelper" Version="30.0.1" />
+		<PackageReference Include="Dapper" Version="2.1.28" />
+		<PackageReference Include="FastMember" Version="1.5.0" />
+		<PackageReference Include="Flurl.Http" Version="4.0.1" />
+		<PackageReference Include="MathNet.Numerics" Version="5.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Identity.Client" Version="4.58.1" />
+		<PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" Version="4.58.1" />
+		<PackageReference Include="Microsoft.IdentityModel.Clients.ActiveDirectory" Version="5.3.0" />
+		<PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
+		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+		<PackageReference Include="SmartFormat.NET" Version="3.3.0" />
+		<PackageReference Include="System.Data.SqlClient" Version="4.8.6" />
+		<!--<PackageReference Include="Microsoft.Windows.CsWinRT" Version="2.0.4" />
+		<PackageReference Include="Microsoft.VisualStudio.Services.InteractiveClient" Version="16.205.1" />
+		<PackageReference Include="WindowsBase" Version="4.6.1055" />-->
+	</ItemGroup>
 
-  <!--<PropertyGroup>
-    <PublishTrimmed>true</PublishTrimmed>
-    <PublishReadyToRun>true</PublishReadyToRun>
-    <PublishSingleFile>true</PublishSingleFile>
-    <RuntimeIdentifier>win-x64</RuntimeIdentifier>
-  </PropertyGroup>-->
+	<ItemGroup>
+		<ProjectReference Include="..\Perfx.Core\Perfx.Core.csproj" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Alba.CsConsoleFormat.ColorfulConsole-NoXaml" Version="1.0.0" />
-    <PackageReference Include="ByteSize" Version="2.0.0" />
-    <PackageReference Include="ClosedXML" Version="0.95.3" />
-    <PackageReference Include="ColoredConsole" Version="1.0.0" />
-    <PackageReference Include="CsvHelper" Version="17.0.0" />
-    <PackageReference Include="Dapper" Version="2.0.78" />
-    <PackageReference Include="FastMember" Version="1.5.0" />
-    <PackageReference Include="Flurl.Http" Version="3.0.0" />
-    <PackageReference Include="MathNet.Numerics" Version="4.12.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Identity.Client" Version="4.23.0" />
-    <PackageReference Include="Microsoft.IdentityModel.Clients.ActiveDirectory" Version="5.2.8" />
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.Toolkit.Wpf.UI.Controls.WebView" Version="6.1.2" />
-    <PackageReference Include="Microsoft.VisualStudio.Services.InteractiveClient" Version="16.153.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-    <PackageReference Include="SmartFormat.NET" Version="2.5.3" />
-    <PackageReference Include="WindowsBase" Version="4.6.1055" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Perfx.Core\Perfx.Core.csproj" />
-  </ItemGroup>
-
-  <!--https://github.com/NuGet/Home/issues/7001-->
-  <Target Name="PostBuild" AfterTargets="PostBuildEvent">
-    <Exec Command="powershell -Command &quot;&amp; {Get-ChildItem -Path '.\bin\Debug\netcoreapp3.1' -File -Recurse | foreach-object {$_.LastWriteTime = (Get-Date)}}&quot;" />
-  </Target>
+	<!--https://github.com/NuGet/Home/issues/7001-->
+	<Target Name="PostBuild" AfterTargets="PostBuildEvent">
+		<Exec Command="powershell -Command &quot;&amp; {Get-ChildItem -Path '.\bin\Debug\net8.0' -File -Recurse | foreach-object {$_.LastWriteTime = (Get-Date)}}&quot;" />
+	</Target>
 
 </Project>
